add a option for separate or no the game sessions on graphic #20
