+++ Services/GameActivityUI.cs
@@ -18,7 +18,7 @@ using System.Windows.Controls;
 using System.Windows.Controls.Primitives;
 using System.Windows.Media;
 
-namespace GameActivity.Views
+namespace GameActivity.Services
 {
     public class GameActivityUI : PlayniteUiHelper
     {
