+++ Views/GameActivitySettingsView.xaml
@@ -21,6 +21,9 @@
                         <CheckBox IsChecked="{Binding showLauncherIcons}" Margin="0,15,0,0">
                             <Label Content="{DynamicResource LOCGameActivityActiveIcon}" />
                         </CheckBox>
+                        <CheckBox IsChecked="{Binding CumulPlaytimeSession}" Margin="0,15,0,0">
+                            <Label Content="{DynamicResource LOCGameActivityCumulPlaytimeSession}" />
+                        </CheckBox>
                     </StackPanel>
 
                     <DockPanel DockPanel.Dock="Bottom" Margin="0,15,0,0">
